split GUI code into submodules
